adjust post components
