+++ src/components/CoursePage/CourseInfo.tsx
